Create working SimpleDiceGame - pure code approach
- Single script creates entire game programmatically
- Click anywhere to roll dice with physics forces
- No GUI setup needed - just attach script and play
- Creates room, walls, dice, camera all in code
- Bypasses Unity GUI workflow completely

Dice roll on click! Next: tune physics forces.

+++ Dice Rolling App/Assets/Scripts/Utils/PerformanceMonitor.cs
@@ -122,7 +122,7 @@ namespace DiceGame.Utils
             currentFPS = 1.0f / Time.unscaledDeltaTime;
             
             // Get memory usage
-            currentMemoryMB = System.GC.GetTotalMemory(false) / (1024 * 1024);
+            currentMemoryMB = (long)(System.GC.GetTotalMemory(false) / (1024 * 1024));
             
             // Update history
             UpdateFPSHistory(currentFPS);
@@ -239,7 +239,7 @@ namespace DiceGame.Utils
         
         public void SetMemoryLimit(long memoryMB)
         {
-            maxMemoryMB = Mathf.Max(50, memoryMB);
+            maxMemoryMB = (long)Mathf.Max(50, memoryMB);
         }
         
         public void ForceQualityReduction()
